+++ dslink-core/src/main/java/com/acuity/iot/dsa/dslink/protocol/protocol_v1/DS1ConnectionInit.java
@@ -31,7 +31,8 @@ public class DS1ConnectionInit extends DSNode {
 
     private static final String DSA_VERSION = "1.1.2";
     //private static final String[] SUPPORTED_FORMATS = new String[]{"msgpack", "json"};
-    private static final String[] SUPPORTED_FORMATS = new String[]{"json"};
+    //private static final String[] SUPPORTED_FORMATS = new String[]{"json"};
+    private static final String[] SUPPORTED_FORMATS = new String[]{"msgpack"};
 
     private String BROKER_REQ = "Broker Request";
     private String BROKER_RES = "Broker Response";
@@ -178,7 +179,7 @@ public class DS1ConnectionInit extends DSNode {
                 buf.append("&token=").append(authToken);
             }
             buf.append("&dsId=").append(getLink().getDsId());
-            buf.append("&format=").append("json");
+            buf.append("&format=").append(response.getString("format"));
         } catch (Exception x) {
             DSException.throwRuntime(x);
         }
@@ -263,9 +264,5 @@ public class DS1ConnectionInit extends DSNode {
             }
         }
     }
-
-    ///////////////////////////////////////////////////////////////////////////
-    // Inner Classes
-    ///////////////////////////////////////////////////////////////////////////
 
 }
